# internal/handlers/spec/api.yaml
openapi: 3.0.0
...
info:
  title: Swagger olidesk API
  description: This is documentation for the Olidesk API.
  contact: {}
  version: 3.0.0
  license:
    name: ""
servers:
  - url: /api
    description: ""
security:
  - BearerAuth: []
paths:
...
  /v1/users/login:
    post:
      tags:
        - Users
      summary: Login user
      description: Login a user
      operationId: postLoginUser
      requestBody:
        description: User data
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginReq"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginRes"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      x-codegen-request-body-name: request
      x-stoplight:
        id: 17ro5fu530gx5
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LoginReq:
      type: object
      properties:
        email:
          type: string
          description: Email de contato
          example: contato@sperium.net
          format: email
          maxLength: 254
          x-go-extra-tags:
            validate: "required,email"
        password:
          type: string
          description: Senha do usuário
          format: password
          minLength: 8
          maxLength: 64
          pattern: "^(?=.*[A-Za-z])(?=.*\\d)(?=.*[^A-Za-z\\d]).{8,64}$"
          x-go-extra-tags:
            validate: "required,min=8,max=64"
      required:
        - email
        - password
      x-stoplight:
        id: rxj62f0yqy7bw
...
    LoginRes:
      type: object
      properties:
        access_token:
          type: string
          description: JWT de acesso
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        token_type:
          type: string
          example: "Bearer"
        expires_in:
          type: integer
          description: Tempo de expiração em segundos
          example: 3600
      required:
        - access_token
        - token_type
      x-stoplight:
        id: lypfoakamq7ky
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
      x-stoplight:
        id: 4qyss6rpqies8